% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_add_dbhydro_dbkeys.R
\name{db_add_dbhydro_dbkeys}
\alias{db_add_dbhydro_dbkeys}
\title{Add DBHYDRO DBKEYs to database}
\usage{
db_add_dbhydro_dbkeys(con, dbkeys)
}
\arguments{
\item{con}{database connection}

\item{dbkeys}{character vector of DBKEY(s)}
}
\value{
logical indicating if operation was successful
}
\description{
Given a set of dbkeys, identifies which dbkey(s) do not
exist in the database, then fetches the metadata for those dbkeys
from DBHYDRO and inserts the results to the dabatase. Also
adds corresponding stations for new dbkeys, if they don't already
exist.
}
\examples{
\dontrun{
db_add_dbhydro_dbkeys(con, c("91346", "91347"))
}
}
