% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_add_dbhydro_stations.R
\name{db_add_dbhydro_stations}
\alias{db_add_dbhydro_stations}
\title{Add DBHYDRO stations to database}
\usage{
db_add_dbhydro_stations(con, station_ids)
}
\arguments{
\item{con}{database connection}

\item{station_ids}{character vector of station_id(s)}
}
\value{
logical indicating if operation was successful
}
\description{
Given a set of station_ids, identifies which station_ids do not
exist in the database, then fetches the metadata for those stations
from DBHYDRO and inserts the results to the dabatase.
}
\examples{
\dontrun{
db_add_dbhydro_stations(con, c("LOX2", "LOX3"))
}
}
