% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connect.R
\name{db_connect}
\alias{db_connect}
\title{Connect to database}
\usage{
db_connect(host, dbname, user, password = NULL, port = 5432)
}
\arguments{
\item{host}{hostname}

\item{dbname}{database name}

\item{user}{username}

\item{password}{password (optional if .pgpass is configured)}

\item{port}{port (default 5432)}
}
\value{
DBI connection object
}
\description{
Connect to database
}
\examples{
\dontrun{
con <- db_connect("localhost", "mydb", "joe", "mypassword")
}
}
