% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_disconnect.R
\name{db_disconnect}
\alias{db_disconnect}
\title{Disconnect from database}
\usage{
db_disconnect(con)
}
\arguments{
\item{con}{connection object}
}
\value{
TRUE if successfull
}
\description{
Disconnect from database
}
\examples{
\dontrun{
con <- db_connect("localhost", "mydb", "joe", "mypassword")
db_disconnect(con)
}
}
