% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_get_dbhydro_dbkeys.R
\name{db_get_dbhydro_dbkeys}
\alias{db_get_dbhydro_dbkeys}
\title{Get DBHYDRO DBKEYs from database}
\usage{
db_get_dbhydro_dbkeys(con, dbkeys = NULL, include_stations = FALSE)
}
\arguments{
\item{con}{database connection}

\item{dbkeys}{character vector of DBKEY(s) to retrieve, or NULL to retrieve all (default)}

\item{include_stations}{include station metadata (join dbhydro_stations table)}
}
\value{
data frame
}
\description{
Get DBHYDRO DBKEYs from database
}
\examples{
\dontrun{
db_get_dbhydro_dbkeys(con) # get all DBKEYs
db_get_dbhydro_dbkeys(con, dbkeys = c("91346", "91347"))
db_get_dbhydro_dbkeys(
  con,
  dbkeys = c("91346", "91347"),
  include_stations = TRUE
) # include station metadata
}
}
