% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_get_dbhydro_hydro.R
\name{db_get_dbhydro_hydro}
\alias{db_get_dbhydro_hydro}
\title{Fetch DBHYDRO hydrologic data from database}
\usage{
db_get_dbhydro_hydro(con, dbkeys, date_min = NULL, date_max = NULL)
}
\arguments{
\item{con}{database connection object}

\item{dbkeys}{character vector of dbkeys}

\item{date_min}{start date (optional)}

\item{date_max}{end date (optional)}
}
\value{
tibble containing hydrologic data
}
\description{
Fetch DBHYDRO hydrologic data from database
}
\examples{
\dontrun{
db_get_dbhydro_hydro(con, dbkeys = c("91473", "91599"))
db_get_dbhydro_hydro(
  con,
  dbkeys = c("91473", "91599"),
  date_min = "2019-10-01",
  date_max = "2019-11-30"
) # specific date range
}
}
