% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_get_dbhydro_stations.R
\name{db_get_dbhydro_stations}
\alias{db_get_dbhydro_stations}
\title{Get DBHYDRO stations from database}
\usage{
db_get_dbhydro_stations(con, station_ids = NULL)
}
\arguments{
\item{con}{database connection}

\item{station_ids}{character vector of station_id(s) to retrieve, or NULL to retrieve all (default)}
}
\value{
data frame
}
\description{
Get DBHYDRO stations from database
}
\examples{
\dontrun{
db_get_dbhydro_stations(con) # get all stations
db_get_dbhydro_stations(con, station_ids = c("LOX3", "LOX6"))
}
}
