% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_get_dbhydro_wq.R
\name{db_get_dbhydro_wq}
\alias{db_get_dbhydro_wq}
\title{Fetch DBHYDRO water quality data from database}
\usage{
db_get_dbhydro_wq(
  con,
  station_ids,
  wq_params = NULL,
  date_min = NULL,
  date_max = NULL
)
}
\arguments{
\item{con}{database connection object}

\item{station_ids}{character vector of dbkeys}

\item{wq_params}{character vector of water quality parameters (optional)}

\item{date_min}{start date (optional)}

\item{date_max}{end date (optional)}
}
\value{
tibble containing water quality data
}
\description{
Fetch DBHYDRO water quality data from database
}
\examples{
\dontrun{
db_get_dbhydro_wq(con, station_ids = c("LOX3", "LOX6"))
db_get_dbhydro_wq(
  con,
  station_ids = c("LOX3", "LOX6"),
  wq_params = "TP",
  date_min = "2019-10-01",
  date_max = "2019-11-30"
) # specific date range
}
}
