% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_get_usgs_dv.R
\name{db_get_usgs_dv}
\alias{db_get_usgs_dv}
\title{Fetch USGS daily values data from database}
\usage{
db_get_usgs_dv(con, station_ids, params = NULL, date_min = NULL,
  date_max = NULL)
}
\arguments{
\item{con}{database connection object}

\item{station_ids}{character vector of station_ids}

\item{params}{character vector of parameters}

\item{date_min}{start date (optional)}

\item{date_max}{end date (optional)}
}
\value{
tibble containing usgs data
}
\description{
Fetch USGS daily values data from database
}
\examples{
\dontrun{
db_get_usgs_dv(con, station_ids = c("263180080205001", "263000080120001"))
db_get_usgs_dv(
  con,
  station_ids = c("263180080205001", "263000080120001"),
  params = c("stage"),
  date_min = "2019-10-01",
  date_max = "2019-11-30"
) # specific params, date range
}
}
