% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_get_usgs_stations.R
\name{db_get_usgs_stations}
\alias{db_get_usgs_stations}
\title{Get USGS stations from database}
\usage{
db_get_usgs_stations(con, station_ids = NULL)
}
\arguments{
\item{con}{database connection}

\item{station_ids}{character vector of station_id(s) to retrieve, or NULL to retrieve all (default)}
}
\value{
data frame
}
\description{
Get USGS stations from database
}
\examples{
\dontrun{
db_get_usgs_stations(con) # get all stations
db_get_usgs_stations(
  con,
  station_ids = c("263180080205001", "263000080120001")
) # specific stations
}
}
