% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_update_dbhydro_hydro.R
\name{db_update_dbhydro_hydro}
\alias{db_update_dbhydro_hydro}
\title{Update hydrologic data in database}
\usage{
db_update_dbhydro_hydro(con, df)
}
\arguments{
\item{con}{database connection object}

\item{df}{data frame containing hydrologic records to insert/update.
Typically generated by \code{dbhydro_clean_hydro(...)} or \code{dbhydro_get_hydro(..., raw = FALSE)}.}
}
\value{
TRUE if operation was successful
}
\description{
Inserts and updates hydrologic data in the database. Performs
an update on records that already exist in the database but
have a outdated revision date. Inserts any new records that do
not already exist in the database.
}
\examples{
\dontrun{
df <- dbhydro_get_hydro(
  dbkeys = c("91473", "91599"),
  date_min = "2019-10-01",
  date_max = "2019-11-30"
)
db_update_dbhydro_hydro(con, df)
}
}
