% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_update_dbhydro_wq.R
\name{db_update_dbhydro_wq}
\alias{db_update_dbhydro_wq}
\title{Update water quality  data in database}
\usage{
db_update_dbhydro_wq(con, df)
}
\arguments{
\item{con}{database connection object}

\item{df}{data frame containing water quality records to insert/update.
Typically generated by \code{dbhydro_clean_wq(...)} or \code{dbhydro_get_wq(..., raw = FALSE)}.}
}
\value{
TRUE if operation was successful
}
\description{
Inserts and updates water quality data in the database. Performs
an update on records that already exist in the database but
have a different result value. Inserts any new records that do
not already exist in the database.
}
\examples{
\dontrun{
df <- dbhydro_get_wq(
  station_ids = c("LOX3", "LOX6"),
  date_min = "2019-01-01",
  date_max = "2019-11-30"
)
db_update_dbhydro_wq(con, df)
}
}
