% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_update_usace_wq.R
\name{db_update_usace_wq}
\alias{db_update_usace_wq}
\title{Update preliminary water quality dataset}
\usage{
db_update_usace_wq(con, df)
}
\arguments{
\item{con}{database connection object}

\item{df}{data frame containing preliminary wq data}
}
\value{
TRUE if operation was successful
}
\description{
Replaces the usace_wq table with new dataset
}
\examples{
\dontrun{
db_update_usace_wq(con, df)
}
}
