% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_update_usgs_dv.R
\name{db_update_usgs_dv}
\alias{db_update_usgs_dv}
\title{Update USGS daily values data in database}
\usage{
db_update_usgs_dv(con, df)
}
\arguments{
\item{con}{database connection object}

\item{df}{data frame containing dv records to insert/update.
Typically generated by \code{usgs_clean_dv(...)} or \code{usgs_get_dv(..., raw = FALSE)}.}
}
\value{
TRUE if operation was successful
}
\description{
Inserts and updates USGS dv data in the database. Performs
an update on records that already exist in the database but
have a outdated revision date. Inserts any new records that do
not already exist in the database.
}
\examples{
\dontrun{
df <- usgs_get_dv(
  station_ids = "263180080205001",
  param = "stage",
  date_min = "2018-01-01",
  date_max = "2018-02-01"
)
db_update_usgs_dv(con, df)
}
}
