% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_upsert_dbhydro_hydro.R
\name{db_upsert_dbhydro_hydro}
\alias{db_upsert_dbhydro_hydro}
\title{Add/update hydrologic data in database}
\usage{
db_upsert_dbhydro_hydro(con, df)
}
\arguments{
\item{con}{database connection object}

\item{df}{data frame containing hydrologic records to insert/update.
Typically generated by \code{dbhydro_clean_hydro(...)} or \code{dbhydro_get_hydro(..., raw = FALSE)}.}
}
\value{
TRUE if operation was successful
}
\description{
Fetches data from DBHYDRO for given list of dbkeys, then performs
upsert operation to insert/update the results in the database.
Existing records will be updated only if the revision date.
}
\examples{
\dontrun{
df <- dbhydro_get_hydro(
  dbkeys = c("91473", "91599"),
  date_min = "2019-10-01",
  date_max = "2019-11-30"
)
db_upsert_dbhydro_hydro(con, df)
}
}
