% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbhydro_batch_get_wq.R
\name{dbhydro_batch_get_wq}
\alias{dbhydro_batch_get_wq}
\title{Get water quality data from DBHYDRO in batches}
\usage{
dbhydro_batch_get_wq(
  station_ids,
  wq_param,
  date_min,
  date_max,
  batch_size = 5,
  raw = FALSE
)
}
\arguments{
\item{station_ids}{character vector containing one or more station_id's}

\item{wq_param}{water quality parameter (must be listed in \code{dbhydro_wq_params})}

\item{date_min}{start date}

\item{date_max}{end date}

\item{batch_size}{number of stations for each batch}

\item{raw}{if TRUE, return raw results from DBHYDRO, otherwise pass results through \code{dbhydro_clean_wq()} before returning (default)}
}
\value{
tibble containing raw data, or empty tibble (no columns) if no data found
}
\description{
Get water quality data from DBHYDRO for specified station(s) and date range
in batches of specified size
}
\examples{
\dontrun{
dbhydro_batch_get_wq(
  station_ids = c("LOX3", "LOX4", "LOX5"),
  wq_param = "TP",
  date_min = "2019-09-01",
  date_max = "2019-10-31",
  batch_size = 2
)
}
}
