% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbhydro_clean_hydro.R
\name{dbhydro_clean_hydro}
\alias{dbhydro_clean_hydro}
\title{Clean hydrologic dataset from DBHYDRO}
\usage{
dbhydro_clean_hydro(x)
}
\arguments{
\item{x}{tibble returned from dbhydro_get_hydro()}
}
\value{
tibble with primary columns
}
\description{
Clean hydrologic dataset from DBHYDRO
}
\examples{
\dontrun{
x <- dbhydro_get_hydro(dbkeys = "91599", date_min = "2019-10-01", date_max = "2019-10-31")
dbhydro_clean_hydro(x)
}
}
