% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbhydro_clean_wq.R
\name{dbhydro_clean_wq}
\alias{dbhydro_clean_wq}
\title{Clean water quality dataset from DBHYDRO}
\usage{
dbhydro_clean_wq(x)
}
\arguments{
\item{x}{tibble returned from dbhydro_get_wq()}
}
\value{
tibble with cleaned water quality data
}
\description{
Clean water quality dataset from DBHYDRO
}
\examples{
\dontrun{
x <- dbhydro_get_wq(
  station_ids = "LOX3",
  wq_param = "TP",
  date_min = "2019-09-01",
  date_max = "2019-10-31"
)
dbhydro_clean_wq(x)
}
}
