% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbhydro_get_dbkey_metadata.R
\name{dbhydro_get_dbkey_metadata}
\alias{dbhydro_get_dbkey_metadata}
\title{Get DBKEY metadata from DBHYDRO}
\usage{
dbhydro_get_dbkey_metadata(dbkeys)
}
\arguments{
\item{dbkeys}{character vector of DBKEYs}
}
\value{
tibble containing metadata for specified DBKEYs
}
\description{
Get DBKEY metadata from DBHYDRO
}
\examples{
\dontrun{
dbhydro_get_dbkey_metadata(dbkeys = c("KQ646"))
dbhydro_get_dbkey_metadata(dbkeys = c("KQ646", "HD889"))
}
}
