% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbhydro_get_hydro.R
\name{dbhydro_get_hydro}
\alias{dbhydro_get_hydro}
\title{Get hydrologic data from dbhydro}
\usage{
dbhydro_get_hydro(dbkeys, date_min, date_max)
}
\arguments{
\item{dbkeys}{character vector containing one or more dbkeys}

\item{date_min}{start date}

\item{date_max}{end date}
}
\value{
tibble containing raw data, or empty tibble (no columns) if no data found
}
\description{
Get hydrologic data from dbhydro for specified dbkey(s) and date range
}
\examples{
dbhydro_get_hydro(
  dbkeys = "91599",
  date_min = "2019-10-01",
  date_max = "2019-10-31"
)
}
