% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbhydro_get_station_metadata.R
\name{dbhydro_get_station_metadata}
\alias{dbhydro_get_station_metadata}
\title{Get water quality station metadata from DBHYDRO}
\usage{
dbhydro_get_station_metadata(station_ids)
}
\arguments{
\item{station_ids}{character vector of station_id(s)}
}
\value{
tibble of metadata for each station_id
}
\description{
Get water quality station metadata from DBHYDRO
}
\examples{
\dontrun{
dbhydro_get_station_metadata(station_ids = c("L3"))
dbhydro_get_station_metadata(station_ids = c("L3", "G136", "G342A", "G342B", "G342C", "G342D"))
}
}
