% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbhydro_get_wq.R
\name{dbhydro_get_wq}
\alias{dbhydro_get_wq}
\title{Get water quality data from DBHYDRO}
\usage{
dbhydro_get_wq(station_ids, wq_param, date_min, date_max, raw = FALSE)
}
\arguments{
\item{station_ids}{character vector containing one or more station_id's}

\item{wq_param}{water quality parameter (must be listed in \code{dbhydro_wq_params})}

\item{date_min}{start date}

\item{date_max}{end date}

\item{raw}{if TRUE, return raw results from DBHYDRO, otherwise pass results through \code{dbhydro_clean_wq()} before returning (default)}
}
\value{
tibble containing raw data, or empty tibble (no columns) if no data found
}
\description{
Get water quality data from DBHYDRO for specified station(s) and date range
}
\examples{
\dontrun{
dbhydro_get_wq(
  station_ids = "LOX3",
  wq_param = "TP",
  date_min = "2019-01-01",
  date_max = "2019-12-31"
)
}
}
