% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dbhydro_wq_params}
\alias{dbhydro_wq_params}
\title{DBHYDRO WQ Parameters.}
\format{A tibble with two variables: \code{test_name} and
  \code{wq_param}}
\usage{
dbhydro_wq_params
}
\description{
Lookup table for translating DBHYDRO \code{test_name} to
\code{wq_param}, which is more concise.
}
\keyword{datasets}
