% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dms_to_ddeg.R
\name{dms_to_ddeg}
\alias{dms_to_ddeg}
\title{Convert latitude/longitude}
\usage{
dms_to_ddeg(x)
}
\arguments{
\item{x}{numeric vector of latitudes/longitudes in DDMMSS.SS format}
}
\value{
numeric vector in decimal degrees
}
\description{
Converts latitude/longitude from DDMMSS.S to decimal degrees
}
\examples{
dms_to_ddeg(262551.3)
dms_to_ddeg(805649.2)
dms_to_ddeg(c(262551.3, 805649.2))
}
