% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker_add.R
\name{tracker_add}
\alias{tracker_add}
\title{Add tracker to the database}
\usage{
tracker_add(con, id, description, hydro = NULL, wq = NULL,
  replace = FALSE)
}
\arguments{
\item{con}{database connection object}

\item{id}{tracker id (character)}

\item{description}{tracker description}

\item{hydro}{data frame containing columns \code{dbkey}, \code{date_min}, \code{date_max}, or NULL}

\item{wq}{data frame containing columns \code{station_id}, \code{wq_param}, \code{date_min}, \code{date_max}, or NULL}

\item{replace}{replace tracker if it already exists in database}
}
\value{
TRUE if tracker was successfully added to the database
}
\description{
Add tracker to the database
}
\examples{
\dontrun{
tracker_add(con, "test-tracker", "A testing tracker", df_dbkeys, df_wq_stations)
}
}
