% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker_data.R
\name{tracker_data}
\alias{tracker_data}
\title{Fetch tracker data from database}
\usage{
tracker_data(con, id, date_min = NULL, date_max = NULL)
}
\arguments{
\item{con}{database connection object}

\item{id}{tracker ID}

\item{date_min}{start date}

\item{date_max}{end date}
}
\value{
list with elements named \code{tracker}, \code{hydro}, and \code{wq}
}
\description{
Retrieves all hydrologic and water quality
data for specified tracker and period
}
\examples{
\dontrun{
tracker_data(con, id = "test-tracker") # all data
tracker_data(
  con,
  id = "test-tracker",
  date_min = "2018-10-01",
  date_max = "2019-09-30"
) # specific period
}
}
