% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker_get.R
\name{tracker_get}
\alias{tracker_get}
\title{Fetch trackers from database}
\usage{
tracker_get(con, ids = NULL)
}
\arguments{
\item{con}{database object}

\item{ids}{character vector of tracker IDs, or NULL to fetch all trackers}
}
\value{
tibble containing trackers
}
\description{
Fetch trackers from database
}
\examples{
\dontrun{
tracker_get(con) # all trackers
tracker_get(con, ids = c("tracker-1", "tracker-2")) # specific trackers
}
}
