% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker_remove.R
\name{tracker_remove}
\alias{tracker_remove}
\title{Remove tracker from database}
\usage{
tracker_remove(con, id, remove_data = FALSE)
}
\arguments{
\item{con}{database connection object}

\item{id}{tracker ID}

\item{remove_data}{remove all hydro and wq data for tracker}
}
\value{
TRUE if tracker was successfully removed
}
\description{
Remove tracker from database
}
\examples{
\dontrun{
tracker_remove(con, "test-tracker")
}
}
