% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker_update.R
\name{tracker_update}
\alias{tracker_update}
\title{Update hydrologic and water quality data for trackers}
\usage{
tracker_update(con, ids = NULL, date_min = NULL, date_max = NULL,
  dbhydro_hydro_batch_size = 5, dbhydro_wq_batch_size = 5)
}
\arguments{
\item{con}{database connection object}

\item{ids}{character vector of tracker IDs, or NULL to update all trackers}

\item{date_min}{start date for period to update, or NULL to use dbkey start date}

\item{date_max}{end date for period to update, or NULL to use dbkey end date}

\item{dbhydro_hydro_batch_size}{batch size passed to \code{dbhydro_batch_get_hydro()} (# dbkeys per batch)}

\item{dbhydro_wq_batch_size}{batch size passed to \code{dbhydro_batch_get_wq()} (# stations per batch)}
}
\value{
TRUE if operation was successful
}
\description{
Update hydrologic and water quality data for trackers
}
\examples{
\dontrun{
tracker_update(con, ids = "test-tracker", date_min = "2019-11-01", date_max = "2019-12-01")
}
}
