% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker_update_dbhydro_hydro.R
\name{tracker_update_dbhydro_hydro}
\alias{tracker_update_dbhydro_hydro}
\title{Update DBHYDRO hydrologic data for trackers}
\usage{
tracker_update_dbhydro_hydro(
  con,
  ids = NULL,
  date_min = NULL,
  date_max = NULL,
  batch_size = 5
)
}
\arguments{
\item{con}{database connection object}

\item{ids}{character vector of tracker IDs, or NULL to update all trackers}

\item{date_min}{start date for period to update, or NULL to use dbkey start date}

\item{date_max}{end date for period to update, or NULL to use dbkey end date}

\item{batch_size}{batch size passed to \code{dbhydro_batch_get_hydro()} (# dbkeys per batch)}
}
\value{
TRUE if operation was successful
}
\description{
Update DBHYDRO hydrologic data for trackers
}
\examples{
\dontrun{
tracker_update_dbhydro_hydro(con, ids = "test-tracker")
}
}
