% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker_update_usgs_dv.R
\name{tracker_update_usgs_dv}
\alias{tracker_update_usgs_dv}
\title{Update USGS daily values datasets for trackers}
\usage{
tracker_update_usgs_dv(
  con,
  ids = NULL,
  date_min = NULL,
  date_max = NULL,
  batch_size = 5
)
}
\arguments{
\item{con}{database connection object}

\item{ids}{character vector of tracker IDs, or NULL to update all trackers}

\item{date_min}{start date for period to update, or NULL to use dbkey start date}

\item{date_max}{end date for period to update, or NULL to use dbkey end date}

\item{batch_size}{batch size passed to \code{dbhydro_batch_get_wq()} (# stations per batch)}
}
\value{
TRUE if operation was successful
}
\description{
Update USGS daily values datasets for trackers
}
\examples{
\dontrun{
tracker_update_usgs_dv(
  con,
  ids = "test-tracker",
  date_min = "2019-01-01",
  date_max = "2019-12-01"
)
}
}
