% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgs_batch_get_dv.R
\name{usgs_batch_get_dv}
\alias{usgs_batch_get_dv}
\title{Get USGS daily values data in batches}
\usage{
usgs_batch_get_dv(
  station_ids,
  param,
  date_min,
  date_max,
  batch_size = 5,
  raw = FALSE
)
}
\arguments{
\item{station_ids}{character vector containing one or more station IDs}

\item{param}{parameter (must be listed in \code{usgs_params})}

\item{date_min}{start date}

\item{date_max}{end date}

\item{batch_size}{number of stations for each batch}

\item{raw}{if TRUE, return raw results from USGS, otherwise pass results through \code{usgs_clean_dv()} before returning (default)}
}
\value{
tibble containing raw/cleaned data, or empty tibble (no columns) if no data found
}
\description{
Get USGS daily values data in batches
}
\examples{
\dontrun{
usgs_batch_get_dv(
  station_ids = c("263180080205001", "263000080120001"),
  param = "stage",
  date_min = "2018-09-01",
  date_max = "2018-10-31",
  batch_size = 1
)
}
}
