% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgs_clean_dv.R
\name{usgs_clean_dv}
\alias{usgs_clean_dv}
\title{Clean USGS daily values dataset}
\usage{
usgs_clean_dv(x)
}
\arguments{
\item{x}{data frame containing raw usgs dataset returned from usgs_get_dv()}
}
\value{
tibble
}
\description{
Clean USGS daily values dataset
}
\examples{
\dontrun{
x <- usgs_get_dv(
  station_ids = "263180080205001",
  param = "STAGE",
  date_min = "2018-01-01",
  date_max = "2018-02-01"
)
usgs_clean_dv(x)
}
}
