% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgs_get_dv.R
\name{usgs_get_dv}
\alias{usgs_get_dv}
\title{Get daily values dataset from USGS NWIS}
\usage{
usgs_get_dv(station_ids, param, date_min, date_max, stat_code = "00003",
  raw = FALSE)
}
\arguments{
\item{station_ids}{character vector of station ID(s)}

\item{param}{parameter name (must exist in usgs_params$param)}

\item{date_min}{start date}

\item{date_max}{end date}

\item{stat_code}{statistic code (default 00003 = daily mean)}

\item{raw}{returns raw dataset returned from dataRetrieval::readNWISdv() if True, otherwise applies usgs_clean_dv()}
}
\value{
data frame (if raw = TRUE), otherwise tibble
}
\description{
Get daily values dataset from USGS NWIS
}
\examples{
\dontrun{
usgs_get_dv(
  station_ids = "263180080205001",
  param = "STAGE",
  date_min = "2018-01-01",
  date_max = "2018-02-01"
)
}
}
