% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgs_get_station_metadata.R
\name{usgs_get_station_metadata}
\alias{usgs_get_station_metadata}
\title{Get station metadata from USGS NWIS}
\usage{
usgs_get_station_metadata(station_ids, raw = FALSE)
}
\arguments{
\item{station_ids}{character vector of station IDs}

\item{raw}{return raw response from dataRetrieval::read_waterdata_monitoring_location() if TRUE, otherwise returns tibble with subset of columns}
}
\value{
tibble with columns \code{station_id}, \code{station_name}, \code{latitude}, \code{longitude} (if \code{raw=FALSE})
}
\description{
Get station metadata from USGS NWIS
}
\examples{
\dontrun{
usgs_get_station_metadata(station_ids = "263180080205001")
}
}
