% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_upsert_dbhydro_hydro.R
\name{db_upsert_dbhydro_hydro}
\alias{db_upsert_dbhydro_hydro}
\title{Add/update hydrologic data in database}
\usage{
db_upsert_dbhydro_hydro(con, dbkeys, date_min, date_max, batch_size = 5)
}
\arguments{
\item{con}{database connection object}

\item{dbkeys}{character vector of dbkeys}

\item{date_min}{start date}

\item{date_max}{end date}

\item{batch_size}{batch size for retrieving data from dbhydro (passed to \code{dbhydro_batch_get_hydro()})}
}
\value{
TRUE if operation was successful
}
\description{
Fetches data from DBHYDRO for given list of dbkeys, then performs
upsert operation to insert/update the results in the database.
Existing records will be updated only if the revision date.
}
\examples{
\dontrun{
db_upsert_dbhydro_hydro(
  con,
  dbkeys = c("91346", "91347"),
  date_min = "2019-10-01",
  date_max = "2019-11-30"
)
}
}
