% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbhydro_get_dbkey_metadata.R
\name{dbhydro_get_dbkey_metadata}
\alias{dbhydro_get_dbkey_metadata}
\title{Get DBKEY metadata from DBHYDRO}
\usage{
dbhydro_get_dbkey_metadata(dbkeys, query = list(), batch_size = 10)
}
\arguments{
\item{dbkeys}{character vector of DBKEYs}

\item{query}{list containing additional query parameters to pass with request}

\item{batch_size}{number of DBKEYs per batch (API limit)}
}
\value{
tibble containing metadata for specified DBKEYs
}
\description{
Get DBKEY metadata from DBHYDRO
}
\examples{
\dontrun{
dbhydro_get_dbkey_metadata(dbkeys = c("KQ646"))
dbhydro_get_dbkey_metadata(dbkeys = c("KQ646", "HD889"))
}
}
