% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbhydro_get_hydro.R
\name{dbhydro_get_hydro}
\alias{dbhydro_get_hydro}
\title{Get hydrologic data from DBHYDRO}
\usage{
dbhydro_get_hydro(dbkeys, date_min, date_max, raw = FALSE)
}
\arguments{
\item{dbkeys}{character vector of DBKEY(s)}

\item{date_min}{start date}

\item{date_max}{end date}

\item{raw}{if TRUE, return raw results from DBHYDRO, otherwise pass results through \code{dbhydro_clean_hydro()} before returning (default)}
}
\value{
tibble containing raw data, or empty tibble (no columns) if no data found
}
\description{
Get hydrologic data from DBHYDRO for specified DBKEY(s) and date range
}
\examples{
\dontrun{
dbhydro_get_hydro(dbkeys = "91599", date_min = "2019-10-01", date_max = "2019-10-31")
}
}
