% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbhydro_get_wq.R
\name{dbhydro_get_wq}
\alias{dbhydro_get_wq}
\title{Get water quality data from dbhydro}
\usage{
dbhydro_get_wq(station_ids, date_min, date_max,
  test_name = "PHOSPHATE, TOTAL AS P")
}
\arguments{
\item{station_ids}{character vector containing one or more station_id's}

\item{date_min}{start date}

\item{date_max}{end date}

\item{test_name}{test name}
}
\value{
tibble containing raw data, or empty tibble (no columns) if no data found
}
\description{
Get water quality data from dbhydro for specified station(s) and date range
}
\examples{
dbhydro_get_wq(
  station_ids = "LOX3",
  date_min = "2019-09-01",
  date_max = "2019-10-31",
  test_name = "PHOSPHATE, TOTAL AS P"
)
}
